feat(Chat): ChatRoom 타입 변경 - unReadMesseges, lastChat 추가 - 연관 컴포넌트 수정
