+++ src/components/Chat/LocalChatRoomList.tsx
@@ -1,20 +1,9 @@
 import { useState } from 'react';
 import { mockChatRoom } from '../../mocks/mockChatRooms';
+import { ChatRoomInfo } from './Chat.type';
 import LocalChatRoom from './LocalChatRoom';
 
-export interface ChatRoomInfo {
-  roomName: string;
-  roomKey: string;
-  description: string;
-  participants: number;
-  maxParticipants: number;
-  lastChat: string;
-  walkingTime: string[];
-}
-
-type Props = {};
-
-function LocalChatRoomList({}: Props) {
+function LocalChatRoomList() {
   const [chatRooms, setChatRooms] = useState<ChatRoomInfo[]>(mockChatRoom);
 
   return (
